first commit
Augmented Book
The aim of this study is to find and analyse any book of academic nature prepared in the field of education and augment any data such as video, audio or some textual information on the given target.
It will fully change the user interaction of reading a book, he will be able to see videos related to the content he will focus through the app's camera augmented on the same content.